feat: melhorias nas Controllers, Repository e Services para utilizar Long no id e injeção de dependência pelo construtor

+++ goodfood-backend/src/main/java/com/goodfood/goodfoodbackend/repositories/NutritionistRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.goodfood.goodfoodbackend.models.Nutritionist;
 
-public interface NutritionistRepository extends JpaRepository <Nutritionist,Integer> {
+public interface NutritionistRepository extends JpaRepository <Nutritionist, Long> {
     
 }
